+++ src/mcplot/color/sron_palettes.py
@@ -17,11 +17,84 @@ History
    * Written Nov 2021 from code *tol_colors.py*
      of Paul Tol, Matthias Cuntz
    * Use : instead of _ in colormap names, Oct 2024, Matthias Cuntz
+   * Add sron:pale and sron:dark cmaps, Oct 2024, Matthias Cuntz
+   * Add sron_named_colors, Oct 2024, Matthias Cuntz
 
 """
 
 
-__all__ = ['sron_colors', 'sron_colormaps', 'sron_functions']
+__all__ = ['sron_named_colors',
+           'sron_colors', 'sron_colormaps', 'sron_functions']
+
+
+sron_named_colors = {
+    # All references to sron_colourschemes.pdf
+    # Fig. 1
+    'sron:bright:blue': '#4477AA',
+    'sron:bright:cyan': '#66CCEE',
+    'sron:bright:green': '#228833',
+    'sron:bright:yellow': '#CCBB44',
+    'sron:bright:red': '#EE6677',
+    'sron:bright:purple': '#AA3377',
+    'sron:bright:grey': '#BBBBBB',
+    # Fig. 2
+    'sron:high-contrast:white': '#FFFFFF',
+    'sron:high-contrast:yellow': '#DDAA33',
+    'sron:high-contrast:red': '#BB5566',
+    'sron:high-contrast:blue': '#004488',
+    'sron:high-contrast:black': '#000000',
+    # Fig. 3
+    'sron:vibrant:blue': '#0077BB',
+    'sron:vibrant:cyan': '#33BBEE',
+    'sron:vibrant:teal': '#009988',
+    'sron:vibrant:orange': '#EE7733',
+    'sron:vibrant:red': '#CC3311',
+    'sron:vibrant:magenta': '#EE3377',
+    'sron:vibrant:grey': '#BBBBBB',
+    # Fig. 4
+    'sron:muted:indigo': '#332288',
+    'sron:muted:cyan': '#88CCEE',
+    'sron:muted:teal': '#44AA99',
+    'sron:muted:green': '#117733',
+    'sron:muted:olive': '#999933',
+    'sron:muted:sand': '#DDCC77',
+    'sron:muted:rose': '#CC6677',
+    'sron:muted:wine': '#882255',
+    'sron:muted:purple': '#AA4499',
+    'sron:muted:grey': '#DDDDDD',
+    # Fig. 5
+    'sron:medium-contrast:white': '#FFFFFF',
+    'sron:medium-contrast:light yellow': '#EECC66',
+    'sron:medium-contrast:light red': '#EE99AA',
+    'sron:medium-contrast:light blue': '#6699CC',
+    'sron:medium-contrast:dark yellow': '#997700',
+    'sron:medium-contrast:dark red': '#994455',
+    'sron:medium-contrast:dark blue': '#004488',
+    'sron:medium-contrast:black': '#000000',
+    # Fig. 6
+    'sron:pale:blue': '#BBCCEE',
+    'sron:pale:cyan': '#CCEEFF',
+    'sron:pale:green': '#CCDDAA',
+    'sron:pale:yellow': '#EEEEBB',
+    'sron:pale:red': '#FFCCCC',
+    'sron:pale:grey': '#DDDDDD',
+    'sron:dark:blue': '#222255',
+    'sron:dark:cyan': '#225555',
+    'sron:dark:green': '#225522',
+    'sron:dark:yellow': '#666633',
+    'sron:dark:red': '#663333',
+    'sron:dark:grey': '#555555',
+    # Fig. 7
+    'sron:light:blue': '#77AADD',
+    'sron:light:cyan': '#99DDFF',
+    'sron:light:mint': '#44BB99',
+    'sron:light:pear': '#BBCC33',
+    'sron:light:olive': '#AAAA00',
+    'sron:light:yellow': '#EEDD88',
+    'sron:light:orange': '#EE8866',
+    'sron:light:pink': '#FFAABB',
+    'sron:light:grey': '#DDDDDD',
+}
 
 
 sron_colors = {
@@ -41,6 +114,11 @@ sron_colors = {
     # Fig. 5 - but order given below Fig. 7
     'sron:medium-contrast': ['#6699CC', '#004488', '#EECC66', '#994455',
                              '#997700', '#EE99AA', '#000000'],
+    # Fig. 6
+    'sron:pale': ['#BBCCEE', '#CCEEFF', '#CCDDAA', '#EEEEBB', '#FFCCCC',
+                  '#DDDDDD'],
+    'sron:dark': ['#222255', '#225555', '#225522', '#666633', '#663333',
+                  '#555555'],
     # Fig. 7 - but order given below Fig. 7
     'sron:light': ['#77AADD', '#EE8866', '#EEDD88', '#FFAABB', '#99DDFF',
                    '#44BB99', '#BBCC33', '#AAAA00', '#DDDDDD', '#000000'],